requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
mysql-connector-python==8.3.0
bcrypt==4.1.3
matplotlib==3.8.0

